chore: move legacy component into /dashboards
